% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sorting_functions.R
\name{get_burden}
\alias{get_burden}
\title{get_burden function}
\usage{
get_burden(disease = "lung", gender = "male", year_range = 2004:2011,
  mortality_data, incident_data, input_list = TRUE, population_std,
  age_average_at_death, input_age_at_onset, standard_LE, standard_LT_age,
  input_duration, input_DisabilityWeight, input_Duration_interval,
  input_DisabilityWeight_interval, age_label = c("0", "1-4", "5-9", "10-14",
  "15-19", "20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54",
  "55-59", "60-64", "65-69", "70-74", "75-79", "80-84", "85+"),
  uncertainty_alpha = 0.05, prior_population = 0.001, nTrials = 2000,
  YLD_perUnit = 1000, personunits = 1000, RateUNIT = 1000, Rate = 0.03,
  Beta = 0.04, Const = 0.1658, Agewt = 0, max_age_YLL = 100,
  max_age_YLD = 90, filepath = paste0(getwd(), "/BOD/R"))
}
\arguments{
\item{disease}{a character indicator of which disease to extract, use when input_list=TRUE, the disease names can be found in description.}

\item{gender}{a character indicator for gender.}

\item{year_range}{a vector showing the range of years.}

\item{mortality_data}{a list of diseases or a dataframe of disease. Each disease is represented in tabular data with row by age and column by year. It requires input of mortality cases.}

\item{incident_data}{a list of diseases or a dataframe of disease. Each disease is represented in tabular data with row by age and column by year. It requires input of incidence cases.}

\item{input_list}{indicates whether mortality_data and incident_data is a list by itself, the example dataset is a list sorted by cancer names, so the (default)input_list=TRUE.}

\item{population_std}{a dataframe of population, with row by age and column by year.}
}
\value{
A list of tabular YLL, YLD, DALY and their rates.
}
\description{
This function estimates YLL, incident-YLD and DALY and their rates
based on input of mortality and incident cases.
Rates are calculated based on input of argument population_std. 
YLL also requires input of age_average_at_death and standard lifetable.
YLD also requires input of input_age_at_onset, input_duration, input_Duration_interval,
input_DisabilityWeight and input_DisabilityWeight_interval.
Current default lifetable is Coale and Demeny's West level 26.
Mortality and incident cases should be in tabular form, row by age and column by year.
The default age group should be 19 groups, starting from 0~1, 1~4, and then by 5 years up to 85+.
}
\examples{
get_burden(disease='lung',
                    year_range=2004:2011,
                    gender='female',
                    mortality_data=mortality_cases_bycause_female,
                    incident_data=incident_cases_bycause_female,
				           input_list=TRUE,
                    population_std=population_std,
                    age_average_at_death=age_average_at_death,
                    input_age_at_onset=age_at_onset,
                    standard_LE=WHO_LE_table[,2],
                    standard_LT_age=WHO_LE_table[,1],  
                    input_duration=input_duration[,1],
                    input_Duration_interval=input_duration[,2:3],
                    input_DisabilityWeight=input_DisabilityWeight,
                    input_DisabilityWeight_interval=input_DisabilityWeight_interval)
}
\keyword{incident_data,}
\keyword{mortality_data}
