% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specific_agegroup_burden.R
\name{get_specific_agegroup_burden}
\alias{get_specific_agegroup_burden}
\title{get_specific_agegroup_burden() function}
\usage{
get_specific_agegroup_burden(data, target_age_group_row = list(c(6:10),
  c(11:14), c(15:19)), target_age_group_lab = c("20-44", "45-64", "65-89"),
  perunit = 1000, digits = 2)
}
\arguments{
\item{data}{input a burden_object object produced by get_burden() function}

\item{target_age_group_row}{it indicates which rows to be combined}

\item{target_age_group_lab}{it indentifies the combined rows with human-readable labels}
}
\value{
A list of tabular YLL, YLD, DALY and their rates, with re-grouped age.
}
\description{
A helper function to sort the data into desired number of age groups.
The function can only take in burden_object produced by get_burden() function.
}
\keyword{data,target_age_group_row,target_age_group_lab}
