% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specific_agegroup_burden.R
\name{get_specific_agegroup_cases}
\alias{get_specific_agegroup_cases}
\title{get_specific_agegroup_cases() function}
\usage{
get_specific_agegroup_cases(cases_data, population_data,
  population_weighted_data, target_age_group_row = list(c(6:10),
  c(11:14), c(15:19)), desired_year = 2004:2011, age_adjusted = TRUE,
  target_age_group_lab = c("20-44", "45-64", "65-89"), perunit = 1000,
  digits = 2)
}
\arguments{
\item{cases_data}{input a dataframe of cases (mortality, incidence or prevalence, all in tabular form).}

\item{population_data}{input a dataframe of population, this population is the local population, for example, a study conducted in Guangzhou, than the Guangzhou population would be used as the population_data.}

\item{population_weighted_data}{input a dataframe of standardized population. Input of this population depends on study setting, in China, the national population or the world population is often used as a standard population.}

\item{target_age_group_row}{it indicates which rows to be combined.}

\item{desired_year}{indicates which periods of years to extract from the cases_data}

\item{target_age_group_lab}{it indentifies the combined rows with human-readable labels.}
}
\value{
A list of tabular cases and case rates, with re-grouped age.
}
\description{
A helper function to sort the cases into desired number of age groups.
The function can only take in dataframe of cases (incidence, prevalence or mortality).
}
\keyword{cases_data,population_data,population_weighted_data}
