% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specific_agegroup_burden.R
\name{get_specific_Yeargroup_cases}
\alias{get_specific_Yeargroup_cases}
\title{get_specific_agegroup_cases() function}
\usage{
get_specific_Yeargroup_cases(data, target_year_col = list(c(1:3), c(4:6),
  c(7:8)), target_year_col_lab = c("2004_2006", "2007_2009",
  "2010_2011"), age_adjusted = TRUE, perunit = 1000, digits = 2)
}
\arguments{
\item{data}{input a reAge_datatable_object object produced by get_specific_agegroup_cases() function}

\item{target_year_col}{it indicates which columns to be combined}

\item{target_year_col_lab}{it indentifies the combined columns with human-readable labels}
}
\value{
A list of tabular cases and case rates, with re-grouped age and year.
}
\description{
A helper function to sort the cases into desired number of periods of year.
The function can only take in reAge_datatable_object produced by get_specific_agegroup_cases() function.
}
\keyword{data,target_year_col,target_year_col_lab}
