% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sorting_functions.R
\name{recal_rate_forBothGen}
\alias{recal_rate_forBothGen}
\title{recal_rate_forBothGen() function}
\usage{
recal_rate_forBothGen(burden_male, burden_female, output_rate_unit = 1000,
  population)
}
\arguments{
\item{burden_male}{a burden_object produced by get_burden() or get_burden_prev(). Only take in the male results.}

\item{burden_female}{a burden_object produced by get_burden() or get_burden_prev(). Only take in the female results.}

\item{output_rate_unit}{a value indicating the unit of rate.}

\item{population}{a list of population. Only two elements should be in the list, each represents a specific gender. Accepted values can be 'male' or 'female'.}
}
\value{
A list of tabular YLL, YLD, DALY and their rates, rates standardized by input of population.
}
\description{
This function works with objects produced by get_burden() or get_burden_prev().
It recalculates burden rate for both gender, based on input of argument population.
}
\examples{
LC_burden_male<-get_burden(disease='lung',year_range=2004:2011,
                    gender='male',
                    mortality_data=mortality_cases_bycause_male,
                    incident_data=incident_cases_bycause_male,
				   input_list=TRUE,
                    population_std=population_std,
                    age_average_at_death=age_average_at_death,
                    input_age_at_onset=age_at_onset,
                    standard_LE=WHO_LE_table[,2],
                    standard_LT_age=WHO_LE_table[,1],
                    input_duration=input_duration[,1],
                    input_Duration_interval=input_duration[,2:3],
                    input_DisabilityWeight=input_DisabilityWeight,
                    input_DisabilityWeight_interval=input_DisabilityWeight_interval)
LC_burden_female<-get_burden(disease='lung',year_range=2004:2011,
                    gender='female',
                    mortality_data=mortality_cases_bycause_female,
                    incident_data=incident_cases_bycause_female,
				   input_list=TRUE,
                    population_std=population_std,
                    age_average_at_death=age_average_at_death,
                    input_age_at_onset=age_at_onset,
                    standard_LE=WHO_LE_table[,2],
                    standard_LT_age=WHO_LE_table[,1],
                    input_duration=input_duration[,1],
                    input_Duration_interval=input_duration[,2:3],
                    input_DisabilityWeight=input_DisabilityWeight,
                    input_DisabilityWeight_interval=input_DisabilityWeight_interval)
recal_rate_forBothGen(burden_male=LC_burden_male,# takes in a burden_object
                               burden_female=LC_burden_female,# takes in a burden_object
                               output_rate_unit=100000,
                               population=population_ex)
}
