% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_test_YLD.R
\name{testing_prevYLD_convergence}
\alias{testing_prevYLD_convergence}
\title{testing_prevYLD_convergence() function}
\usage{
testing_prevYLD_convergence(iter = 100, data, filepath = paste0(getwd(),
  "/R"))
}
\arguments{
\item{iter}{indicates number of iteration}

\item{data}{must take in a YLD_object produced by YLD_prevalent(), not by YLD_incident()}
}
\value{
A list of Convergence_Test_object
}
\description{
This function perform convergence test for YLD, and output in a dataset full of sampling data
The argument data can only take in a YLD_object produced by YLD_prevalent(), not by YLD_incident()
}
\keyword{data,}
\keyword{iter}
