% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_test.R
\name{testing_SLE_convergence}
\alias{testing_SLE_convergence}
\title{testing_SLE_convergence() function}
\usage{
testing_SLE_convergence(iter = 100, data)
}
\arguments{
\item{iter}{indicates number of iteration}

\item{data}{must take in a SLE_object produced by SLE()}
}
\value{
A list of Convergence_Test_object
}
\description{
This function perform convergence test for SLE, and output in a dataset full of sampling data
The argument data can only take in a SLE_object produced by SLE()
}
\keyword{data,}
\keyword{iter}
