% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_test_YLD.R
\name{testing_YLD_convergence}
\alias{testing_YLD_convergence}
\title{testing_YLD_convergence() function}
\usage{
testing_YLD_convergence(iter = 100, data)
}
\arguments{
\item{iter}{indicates number of iteration}

\item{data}{must take in a YLD_object produced by YLD_incident(), not by YLD_prevalent()}
}
\value{
A list of Convergence_Test_object
}
\description{
This function perform convergence test for YLD, and output in a dataset full of sampling data
The argument data can only take in a YLD_object produced by YLD_incident(), not by YLD_prevalent()
}
\keyword{data,}
\keyword{iter}
